+++ feature_bill/src/main/java/dev/lkey/bill/di/modules/BillUseCaseModule.kt
@@ -5,9 +5,9 @@ import dagger.Provides
 import dev.lkey.account.domain.AccountRepository
 import dev.lkey.bill.domain.repository.BillRepository
 import dev.lkey.bill.domain.usecase.GetBillInfoUseCase
+import dev.lkey.bill.domain.usecase.GetTransactionsUseCase
 import dev.lkey.bill.domain.usecase.UpdateBillUseCase
-import dev.lkey.transations.domain.repository.TransactionsRepository
-import dev.lkey.transations.domain.usecase.GetTransactionsUseCase
+import dev.lkey.transactions.domain.TransactionsDatasource
 
 /**
  * Модуль use-case счета
@@ -32,9 +32,10 @@ class BillUseCaseModule {
 
     @Provides
     fun provideGetTransactionsUseCase(
-        transactionsRepository: TransactionsRepository
+        transactionsDatasource: TransactionsDatasource
     ): GetTransactionsUseCase {
-        return GetTransactionsUseCase(transactionsRepository)
+        return GetTransactionsUseCase(transactionsDatasource)
     }
+
 
 }
