+++ feature_transactions/src/main/java/dev/lkey/transations/domain/repository/TransactionsRepository.kt
@@ -1,15 +1,8 @@
 package dev.lkey.transations.domain.repository
 
-import dev.lkey.common.core.model.transaction.TransactionModel
-import dev.lkey.transations.data.dto.RequestTransactionDto
+import dev.lkey.transactions.data.dto.RequestTransactionDto
 
 interface TransactionsRepository {
-
-    suspend fun getTransactions(
-        accountId : Int,
-        startDate: String,
-        endDate: String
-    ): Result<List<TransactionModel>>
 
     suspend fun createTransaction(
         transaction: RequestTransactionDto
@@ -23,7 +16,5 @@ interface TransactionsRepository {
     suspend fun deleteTransaction(
         id: Int
     ): Result<Unit>
-
-    suspend fun uploadUnsyncedTransactions(): Result<Unit>
 
 }
