+++ feature_transactions/src/main/java/dev/lkey/transations/presentation/create/viewmodel/CreateTransactionViewModel.kt
@@ -9,7 +9,7 @@ import dev.lkey.core.error.ErrorHandler
 import dev.lkey.core.error.OfflineDataException
 import dev.lkey.core.network.FinancilityResult
 import dev.lkey.storage.data.sync.AppSyncStorage
-import dev.lkey.transations.data.dto.RequestTransactionDto
+import dev.lkey.transactions.data.dto.RequestTransactionDto
 import dev.lkey.transations.domain.usecase.GetAccountsUseCase
 import dev.lkey.transations.domain.usecase.GetArticlesUseCase
 import dev.lkey.transations.domain.usecase.PostTransactionUseCase
