+++ settings.gradle.kts
@@ -26,3 +26,4 @@ include(":storage")
 include(":articles")
 include(":account")
 include(":feature_splash")
+include(":transactions")
