+++ transactions/src/main/java/dev/lkey/transactions/data/mappers/toTransactionDto.kt
@@ -1,7 +1,7 @@
-package dev.lkey.transations.data.mappers
+package dev.lkey.transactions.data.mappers
 
 import dev.lkey.storage.data.model.TransactionEntity
-import dev.lkey.transations.data.dto.RequestTransactionDto
+import dev.lkey.transactions.data.dto.RequestTransactionDto
 
 fun TransactionEntity.toTransactionDto(): RequestTransactionDto {
