+++ transactions/src/main/java/dev/lkey/transactions/domain/TransactionsDatasource.kt
@@ -1,0 +1,15 @@
+package dev.lkey.transactions.domain
+
+import dev.lkey.common.core.model.transaction.TransactionModel
+
+interface TransactionsDatasource {
+
+    suspend fun getTransactions(
+        accountId : Int,
+        startDate: String,
+        endDate: String
+    ): Result<List<TransactionModel>>
+
+    suspend fun uploadUnsyncedTransactions(): Result<Unit>
+
+}
